% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_cond.R
\name{mutate_cond}
\alias{mutate_cond}
\title{mutate_cond}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{a dataframe or data table}

\item{condition}{an expreessed condition to mutate on}

\item{...}{additional conditions separated by commas}
}
\value{
a mutated dataframe
}
\description{
A function that mutates data frames but only acts on the rows satisfying the condition
}
\examples{
DF \%>\% mutate_cond(measure == 'exit', qty.exit = qty, cf = 0, delta.watts = 13)
}
