% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_last.R
\name{mutate_last}
\alias{mutate_last}
\title{mutate_last}
\usage{
mutate_last(.data, ...)
}
\arguments{
\item{.data}{a dataframe or data table}

\item{...}{conditions or set of conditions to mutate on last/greatest group}
}
\value{
a mutated dataframe or data table
}
\description{
A function like mutate but operates only within a group_by and only operates on the last group rather than every group.
}
\examples{

DF \%>\%
group_by(is.exit = measure == 'exit') \%>\%
  mutate_last(qty.exit = qty, cf = 0, delta.watts = 13) \%>\%
  ungroup() \%>\%
  select(-is.exit)

}
